+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Added mutable param to use inside worker.Util to avoid locking
 - Changed peek on iterator to not be mutable. Now get field, not need a mutable row.
 - Non mutable peek, allow to repair macros get_Str, get string, get_str_utf8
+- Downgrade edition to 2018, to add compatibility.
 
 ---